backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
python-dotenv==1.0.1
pydantic==2.12.5
langchain==0.3.25
langchain-core==0.3.84
langchain-groq==0.2.5
langchain-community==0.3.24
langchain-text-splitters==0.3.8
tavily-python==0.7.23
opik==1.11.2
groq==0.37.1
